+++ src/components/common/ModulePlaceholder.jsx
